src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
bitstring==4.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
discord.py==2.3.2
dnspython==2.6.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
quickchart-io==2.0.0
requests==2.32.2
threadpoolctl==3.5.0
urllib3==2.2.1
yarl==1.9.4
